requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
chromadb
groq
python-dotenv
pydantic
requests
beautifulsoup4
moviepy==1.0.3
pillow